Fix version conflict by vendoring an updated glamor
(See previous commit message for context.)

First, we fixed glamor to work with the current version (post-v5) of
inline-style-prefixer (ISP). For full details on those changes, see this
PR we opened: https://github.com/threepointone/glamor/pull/396

After that, we:

- cloned that (forked) version of glamor to vendor/glamor,
- ran `npm install` in that directory,
- ran `npm run build` in that directory to compile the source and make
  library paths available, and
- removed unneeded files with `rm -rf .git node_modules`.

Finally, we added two new dependencies to our project:

- glamor, pointing to the vendored code, and
- ISP at ^6.0.0, which at this point both ui-box and (our local version
  of) glamor can use without a problem.

(For some reason we don't yet understand, only adding glamor wasn't
enough. When we did that, ISP wasn't getting installed at the top level
of node_modules, and Shadow CLJS wasn't able to find it, so we had to
add it as a top-level dep.)

After that and a top-level `npm install`, we could `npm run watch` and
have a working example that used a previously broken feature of
Evergreen.

Credit: Christopher Gsell for pair programming me on this journey.